// src/lib/contracts/generated/Pool.ts
interface IParams {
  address: string;
  web3: Web3;
  sender?: string; // the address of sender, if undefined then web3.eth.defaultAccount used
  gasEstimationMultiplayer?: number; // if undefined then 1
}

const DEFAULT_GAS_ESTIMATION_MULTIPLAYER = 1.1; // + 10 %

export class Pool {
  native: Web3Contract;
  gasEstimationMultiplayer: number;
  web3: Web3;
  _sender: string | undefined;

  get sender(): string | undefined {
    if (this._sender) return this._sender;
    if (this.web3.defaultAccount) return this.web3.defaultAccount;
    return undefined;
  }

  protected getSenderOrFail(): string {
    const sender = this.sender;
    if (!sender) {
      throw new Error("Sender is required");
    }
    return sender;
  }

  constructor({ address, web3, sender, gasEstimationMultiplayer }: IParams) {
    this._sender = sender;
    this.native = new web3.eth.Contract(Abi, address) as any;
    this.gasEstimationMultiplayer =
      gasEstimationMultiplayer ?? DEFAULT_GAS_ESTIMATION_MULTIPLAYER;
    this.web3 = web3;
  }

  protected async estimateGas<
    K,
    T extends PayableTransactionObject<K> | NonPayableTransactionObject<K>
  >(method: T, args?: Parameters<T["estimateGas"]>) {
    const originalEstimation = await method.estimateGas(...(args ?? []));
    return new BN(originalEstimation)
      .muln(this.gasEstimationMultiplayer)
      .toString();
  }

  async addAdmin(account: string): Promise<PromiEvent<TransactionReceipt>> {
    const method = this.native.methods.addAdmin(account);
    return method.send({
      gas: await this.estimateGas(method),
      from: this.getSenderOrFail()
    });
  }

  async addDao(account: string): Promise<PromiEvent<TransactionReceipt>> {
    const method = this.native.methods.addDao(account);
    return method.send({
      gas: await this.estimateGas(method),
      from: this.getSenderOrFail()
    });
  }

  async isAdmin(account: string): Promise<boolean> {
    return this.native.methods.isAdmin(account).call();
  }

  async isDao(account: string): Promise<boolean> {
    return this.native.methods.isDao(account).call();
  }

  async oneInchAddr(): Promise<string> {
    return this.native.methods.oneInchAddr().call();
  }

  async renounceAdmin(): Promise<PromiEvent<TransactionReceipt>> {
    const method = this.native.methods.renounceAdmin();
    return method.send({
      gas: await this.estimateGas(method),
      from: this.getSenderOrFail()
    });
  }

  async renounceDao(): Promise<PromiEvent<TransactionReceipt>> {
    const method = this.native.methods.renounceDao();
    return method.send({
      gas: await this.estimateGas(method),
      from: this.getSenderOrFail()
    });
  }

  async swap(
    srcToken: string,
    srcAmount: number | string | BN,
    minPrice: number | string | BN,
    maxPrice: number | string | BN,
    destToken: string,
    oneInchData: string | number[]
  ): Promise<PromiEvent<TransactionReceipt>> {
    const method = this.native.methods.swap(
      srcToken,
      srcAmount,
      minPrice,
      maxPrice,
      destToken,
      oneInchData
    );
    return method.send({
      gas: await this.estimateGas(method),
      from: this.getSenderOrFail()
    });
  }

  async transfer(
    token: string,
    receiver: string,
    amount: number | string | BN
  ): Promise<PromiEvent<TransactionReceipt>> {
    const method = this.native.methods.transfer(token, receiver, amount);
    return method.send({
      gas: await this.estimateGas(method),
      from: this.getSenderOrFail()
    });
  }

  async updateOneInchAddress(
    newOneInch: string
  ): Promise<PromiEvent<TransactionReceipt>> {
    const method = this.native.methods.updateOneInchAddress(newOneInch);
    return method.send({
      gas: await this.estimateGas(method),
      from: this.getSenderOrFail()
    });
  }

  async getPoolBalance(asset: string): Promise<string> {
    return this.native.methods.getPoolBalance(asset).call();
  }
}