// src/lib/contracts/generated/Timelock.ts
interface IParams {
  address: string;
  web3: Web3;
  sender?: string; // the address of sender, if undefined then web3.eth.defaultAccount used
  gasEstimationMultiplayer?: number; // if undefined then 1
}

const DEFAULT_GAS_ESTIMATION_MULTIPLAYER = 1.1; // + 10 %

export class Timelock {
  native: Web3Contract;
  gasEstimationMultiplayer: number;
  web3: Web3;
  _sender: string | undefined;

  get sender(): string | undefined {
    if (this._sender) return this._sender;
    if (this.web3.defaultAccount) return this.web3.defaultAccount;
    return undefined;
  }

  protected getSenderOrFail(): string {
    const sender = this.sender;
    if (!sender) {
      throw new Error("Sender is required");
    }
    return sender;
  }

  constructor({ address, web3, sender, gasEstimationMultiplayer }: IParams) {
    this._sender = sender;
    this.native = new web3.eth.Contract(Abi, address) as any;
    this.gasEstimationMultiplayer =
      gasEstimationMultiplayer ?? DEFAULT_GAS_ESTIMATION_MULTIPLAYER;
    this.web3 = web3;
  }

  protected async estimateGas<
    K,
    T extends PayableTransactionObject<K> | NonPayableTransactionObject<K>
  >(method: T, args?: Parameters<T["estimateGas"]>) {
    const originalEstimation = await method.estimateGas(...(args ?? []));
    return new BN(originalEstimation)
      .muln(this.gasEstimationMultiplayer)
      .toString();
  }

  async GRACE_PERIOD(): Promise<string> {
    return this.native.methods.GRACE_PERIOD().call();
  }

  async MAXIMUM_DELAY(): Promise<string> {
    return this.native.methods.MAXIMUM_DELAY().call();
  }

  async MINIMUM_DELAY(): Promise<string> {
    return this.native.methods.MINIMUM_DELAY().call();
  }

  async admin(): Promise<string> {
    return this.native.methods.admin().call();
  }

  async delay(): Promise<string> {
    return this.native.methods.delay().call();
  }

  async pendingAdmin(): Promise<string> {
    return this.native.methods.pendingAdmin().call();
  }

  async queuedTransactions(arg0: string | number[]): Promise<boolean> {
    return this.native.methods.queuedTransactions(arg0).call();
  }

  async setDelay(
    delay_: number | string | BN
  ): Promise<PromiEvent<TransactionReceipt>> {
    const method = this.native.methods.setDelay(delay_);
    return method.send({
      gas: await this.estimateGas(method),
      from: this.getSenderOrFail()
    });
  }

  async acceptAdmin(): Promise<PromiEvent<TransactionReceipt>> {
    const method = this.native.methods.acceptAdmin();
    return method.send({
      gas: await this.estimateGas(method),
      from: this.getSenderOrFail()
    });
  }

  async setPendingAdmin(
    pendingAdmin_: string
  ): Promise<PromiEvent<TransactionReceipt>> {
    const method = this.native.methods.setPendingAdmin(pendingAdmin_);
    return method.send({
      gas: await this.estimateGas(method),
      from: this.getSenderOrFail()
    });
  }

  async queueTransaction(
    target: string,
    value: number | string | BN,
    signature: string,
    data: string | number[],
    eta: number | string | BN
  ): Promise<PromiEvent<TransactionReceipt>> {
    const method = this.native.methods.queueTransaction(
      target,
      value,
      signature,
      data,
      eta
    );
    return method.send({
      gas: await this.estimateGas(method),
      from: this.getSenderOrFail()
    });
  }

  async cancelTransaction(
    target: string,
    value: number | string | BN,
    signature: string,
    data: string | number[],
    eta: number | string | BN
  ): Promise<PromiEvent<TransactionReceipt>> {
    const method = this.native.methods.cancelTransaction(
      target,
      value,
      signature,
      data,
      eta
    );
    return method.send({
      gas: await this.estimateGas(method),
      from: this.getSenderOrFail()
    });
  }

  async executeTransaction(
    target: string,
    value: number | string | BN,
    signature: string,
    data: string | number[],
    eta: number | string | BN
  ): Promise<PromiEvent<TransactionReceipt>> {
    const method = this.native.methods.executeTransaction(
      target,
      value,
      signature,
      data,
      eta
    );
    return method.send({
      gas: await this.estimateGas(method),
      from: this.getSenderOrFail()
    });
  }
}